Make repo installable as a pip package
