+++ autovar/test_comparison.py
@@ -2,9 +2,9 @@ from astropy.io import fits
 import numpy
 import os
 import glob
-from .test_files import output_tests
 
-from .comparison import find_comparisons, read_data_files, find_reference_frame
+from autovar.test_files import output_tests
+from autovar.comparison import find_comparisons, read_data_files, find_reference_frame
 
 TEST_DATA_PATH = os.path.join(os.getcwd(),'test_files')
