+++ autovar/test_files/comparison/Waspd43b_w_30d27_2018d02d14T03d26d06d777_1a2157407_2458163d6390746250_kb95.psx
