+++ autovar/test_files/comparison/Waspd43b_w_30d27_2018d02d14T04d08d18d637_1a1253206_2458163d6683783028_kb95.psx
