+++ autovar/test_files/stars/photometry_test.fits
