+++ autovar/test_files/stars/photometry_test2.fits
