+++ autovar/test_files/stars/usedImages_test.txt
